+++ Editor/translations/Editor_zh_CN.ts
@@ -686,6 +686,25 @@ Select a different file or change the device configuration.</source>
     </message>
 </context>
 <context>
+    <name>DelayCardEditor</name>
+    <message>
+        <location filename="../widgets/cards/DelayCardEditor.cpp" line="70"/>
+        <source>Time</source>
+        <translation type="unfinished">时间</translation>
+    </message>
+    <message>
+        <location filename="../widgets/cards/DelayCardEditor.cpp" line="71"/>
+        <source>Samples</source>
+        <translation type="unfinished">采样</translation>
+    </message>
+    <message>
+        <location filename="../widgets/cards/DelayCardEditor.cpp" line="78"/>
+        <location filename="../widgets/cards/DelayCardEditor.cpp" line="117"/>
+        <source>samples</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>DelayFilterGUI</name>
     <message>
         <location filename="../guis/DelayFilterGUI.ui" line="32"/>
@@ -973,6 +992,44 @@ Select a different file or change the device configuration.</source>
     </message>
 </context>
 <context>
+    <name>ExpressionFilterGUIFactory</name>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="35"/>
+        <source>Control</source>
+        <translation type="unfinished">控制</translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="36"/>
+        <source>Branching</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="39"/>
+        <source>If (Begin conditional section)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="40"/>
+        <source>ElseIf (Alternative condition)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="41"/>
+        <source>Else (Fallback section)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="42"/>
+        <source>EndIf (End conditional section)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../guis/ExpressionFilterGUIFactory.cpp" line="38"/>
+        <source>Eval (Evaluate expression)</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>FilterCardModel</name>
     <message>
         <location filename="../widgets/FilterCardModel.cpp" line="203"/>
@@ -1000,80 +1057,120 @@ Select a different file or change the device configuration.</source>
         <translation>Biquad</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="270"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="261"/>
+        <source>IIR filter</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="266"/>
+        <source>Order %1</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="269"/>
+        <source>%1 coefficients</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="299"/>
         <source>Graphic EQ</source>
         <translation>图形均衡器</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="275"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="304"/>
         <source>%1 bands</source>
         <translation>%1 个频段</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="281"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="310"/>
         <source>Copy</source>
         <translation>复制</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="303"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="332"/>
         <source>%1 steps, %2 virtual</source>
         <translation>%1 项，%2 个虚拟声道</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="305"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="334"/>
         <source>%1 steps</source>
         <translation>%1 项</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="312"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="341"/>
         <source>Channel</source>
         <translation>声道</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="322"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="351"/>
         <source>Include</source>
         <translation>包含</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="333"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="362"/>
         <source>Convolution</source>
         <translation>卷积</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="347"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="376"/>
         <source>MultiConvolution</source>
         <translation>多路卷积</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="362"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="391"/>
         <source>VST Plugin</source>
         <translation>VST 插件</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="372"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="401"/>
         <source>Device</source>
         <translation>设备</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="379"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="408"/>
         <source>Stage</source>
         <translation>阶段</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardModel.cpp" line="386"/>
+        <location filename="../widgets/FilterCardModel.cpp" line="415"/>
         <source>Loudness</source>
         <translation>响度</translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="430"/>
+        <source>If</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="435"/>
+        <source>Else if</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="440"/>
+        <source>Else</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="445"/>
+        <source>End if</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/FilterCardModel.cpp" line="452"/>
+        <source>Eval</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
     <name>FilterCardRow</name>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="60"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="61"/>
         <source>Expand filter card</source>
         <translation>展开滤波器卡片</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="96"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="97"/>
         <source>Enable or comment out this command</source>
         <translation>启用或注释掉此命令</translation>
     </message>
@@ -1082,22 +1179,22 @@ Select a different file or change the device configuration.</source>
         <translation type="vanished">在此卡片前添加滤波器</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="105"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="106"/>
         <source>Add filter after this card</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="112"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="113"/>
         <source>Remove filter</source>
         <translation>移除滤波器</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="120"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="121"/>
         <source>Edit raw command</source>
         <translation>编辑原始命令</translation>
     </message>
     <message>
-        <location filename="../widgets/FilterCardRow.cpp" line="596"/>
+        <location filename="../widgets/FilterCardRow.cpp" line="664"/>
         <source>Raw</source>
         <translation>原始</translation>
     </message>
@@ -1380,6 +1477,24 @@ Select a different file or change the device configuration.</source>
         <location filename="../guis/GraphicEQFilterGUIFactory.cpp" line="41"/>
         <source>Graphic equalizer with variable bands</source>
         <translation>具有可变频段的图形均衡器</translation>
+    </message>
+</context>
+<context>
+    <name>IIRCardEditor</name>
+    <message>
+        <location filename="../widgets/cards/IIRCardEditor.cpp" line="57"/>
+        <source>Order</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/cards/IIRCardEditor.cpp" line="76"/>
+        <source>Feedforward (b)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../widgets/cards/IIRCardEditor.cpp" line="83"/>
+        <source>Feedback (a)</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
@@ -2070,12 +2185,12 @@ Do you want to run the Configurator application to fix the problem?</source>
         <translation>此设备的流格式为 %0（%1 位容器）。EqualizerAPO 仅原生处理 IEEE_FLOAT 32/64 位音频流，因此音频会被转发而不应用任何滤波器。如果你需要对此设备进行滤波处理，请在声音设置中将该设备的默认格式切换为 32 位 IEEE_FLOAT。</translation>
     </message>
     <message>
-        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="104"/>
+        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="129"/>
         <source>%0 dB</source>
         <translation>%0 dB</translation>
     </message>
     <message>
-        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="123"/>
+        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="148"/>
         <source>%0 %</source>
         <translation type="unfinished"></translation>
     </message>
@@ -2084,7 +2199,7 @@ Do you want to run the Configurator application to fix the problem?</source>
         <translation type="vanished">%0 毫秒（%1 个采样）</translation>
     </message>
     <message>
-        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="118"/>
+        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="143"/>
         <source>%0 ms (%1 s.)</source>
         <translation>%0 毫秒（%1 秒）</translation>
     </message>
@@ -2093,8 +2208,8 @@ Do you want to run the Configurator application to fix the problem?</source>
         <translation type="vanished">%0 毫秒（%1 个采样）</translation>
     </message>
     <message>
-        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="113"/>
-        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="120"/>
+        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="138"/>
+        <location filename="../MainWindowParts/MainWindow.Analysis.cpp" line="145"/>
         <source>%0 ms</source>
         <translation>%0 毫秒</translation>
     </message>
@@ -2750,28 +2865,33 @@ Do you want to run the Device Selector application to fix the problem?</source>
 <context>
     <name>SoftFilterPickerView</name>
     <message>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="179"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="193"/>
+        <source>Complete as it is</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="195"/>
         <source>Choose the details after adding</source>
         <translation>添加后再设置详细信息</translation>
     </message>
     <message>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="414"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="430"/>
         <source>Search filters</source>
         <translation>搜索滤波器</translation>
     </message>
     <message>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="453"/>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="514"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="469"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="530"/>
         <source>General</source>
         <translation>常规</translation>
     </message>
     <message>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="562"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="578"/>
         <source>Nothing matches your search</source>
         <translation>没有匹配你搜索的内容</translation>
     </message>
     <message>
-        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="563"/>
+        <location filename="../skins/pickers/SoftFilterPicker.cpp" line="579"/>
         <source>Try a shorter or different keyword</source>
         <translation>请尝试更短或不同的关键词</translation>
     </message>
@@ -2779,7 +2899,82 @@ Do you want to run the Device Selector application to fix the problem?</source>
 <context>
     <name>SoftSkin</name>
     <message>
-        <location filename="../skins/SoftSkin.cpp" line="725"/>
+        <location filename="../skins/SoftSkin.cpp" line="100"/>
+        <source>Otherwise</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="102"/>
+        <source>End of the rule</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="113"/>
+        <source>Set %1 to %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="134"/>
+        <source>If %1 is %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="136"/>
+        <source>If %1 is not %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="138"/>
+        <source>If %1 is at least %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="140"/>
+        <source>If %1 is more than %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="142"/>
+        <source>If %1 is at most %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="144"/>
+        <source>If %1 is less than %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="149"/>
+        <source>Otherwise, if %1 is %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="151"/>
+        <source>Otherwise, if %1 is not %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="153"/>
+        <source>Otherwise, if %1 is at least %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="155"/>
+        <source>Otherwise, if %1 is more than %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="157"/>
+        <source>Otherwise, if %1 is at most %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="159"/>
+        <source>Otherwise, if %1 is less than %2</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../skins/SoftSkin.cpp" line="826"/>
         <source>Sound may distort - keep it below 0 dB</source>
         <translation type="unfinished"></translation>
     </message>